# packages/contracts/openapi/opencore-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/collaboration/summary": {
      "get": {
        "operationId": "CollaborationController_getSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollaborationSummaryDto"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration"]
      }
    },
    "/api/collaboration/messages": {
      "get": {
        "operationId": "CollaborationController_listMessages",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["unread", "read", "archived"]
            }
          },
          {
            "name": "recipient",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessagePageDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Messages"]
      },
      "post": {
        "operationId": "CollaborationController_createMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateMessageDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Messages"]
      }
    },
    "/api/collaboration/messages/{id}": {
      "get": {
        "operationId": "CollaborationController_getMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Messages"]
      },
      "delete": {
        "operationId": "CollaborationController_deleteMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteResultDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Messages"]
      }
    },
    "/api/collaboration/messages/{id}/read": {
      "patch": {
        "operationId": "CollaborationController_markMessageRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Messages"]
      }
    },
    "/api/collaboration/messages/{id}/archive": {
      "patch": {
        "operationId": "CollaborationController_archiveMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Messages"]
      }
    },
    "/api/collaboration/notices": {
      "get": {
        "operationId": "CollaborationController_listNotices",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["draft", "published", "archived"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NoticePageDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Notices"]
      },
      "post": {
        "operationId": "CollaborationController_createNotice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateNoticeDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NoticeDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Notices"]
      }
    },
    "/api/collaboration/notices/{id}": {
      "get": {
        "operationId": "CollaborationController_getNotice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NoticeDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Notices"]
      }
    },
    "/api/collaboration/notices/{id}/publish": {
      "patch": {
        "operationId": "CollaborationController_publishNotice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NoticeDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Notices"]
      }
    },
    "/api/collaboration/notices/{id}/archive": {
      "patch": {
        "operationId": "CollaborationController_archiveNotice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NoticeDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Notices"]
      }
    },
    "/api/collaboration/todos": {
      "get": {
        "operationId": "CollaborationController_listTodos",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pending", "assigned", "completed", "canceled"]
            }
          },
          {
            "name": "assignee",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "sourceType",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoPageDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Todos"]
      },
      "post": {
        "operationId": "CollaborationController_createTodo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTodoDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Todos"]
      }
    },
    "/api/collaboration/todos/{id}": {
      "get": {
        "operationId": "CollaborationController_getTodo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Todos"]
      }
    },
    "/api/collaboration/todos/{id}/assign": {
      "patch": {
        "operationId": "CollaborationController_assignTodo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssignTodoDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Todos"]
      }
    },
    "/api/collaboration/todos/{id}/complete": {
      "patch": {
        "operationId": "CollaborationController_completeTodo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TodoActionDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Todos"]
      }
    },
    "/api/collaboration/todos/{id}/cancel": {
      "patch": {
        "operationId": "CollaborationController_cancelTodo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TodoActionDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Todos"]
      }
    },
    "/api/collaboration/approvals": {
      "get": {
        "operationId": "CollaborationController_listApprovalLiteRequests",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pending", "approved", "rejected"]
            }
          },
          {
            "name": "requester",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "approver",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApprovalLitePageDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Approval Lite"]
      },
      "post": {
        "operationId": "CollaborationController_createApprovalLiteRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateApprovalLiteDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApprovalLiteDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Approval Lite"]
      }
    },
    "/api/collaboration/approvals/{id}": {
      "get": {
        "operationId": "CollaborationController_getApprovalLiteRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApprovalLiteDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Approval Lite"]
      }
    },
    "/api/collaboration/approvals/{id}/approve": {
      "patch": {
        "operationId": "CollaborationController_approveApprovalLiteRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DecideApprovalLiteDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApprovalLiteDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Approval Lite"]
      }
    },
    "/api/collaboration/approvals/{id}/reject": {
      "patch": {
        "operationId": "CollaborationController_rejectApprovalLiteRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DecideApprovalLiteDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApprovalLiteDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Collaboration", "Collaboration Approval Lite"]
      }
    },
...
  },
  "info": {
    "title": "OpenCore API",
    "description": "OpenCore API contract",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [{ "url": "/" }],
  "components": {
    "schemas": {
...
      "DeleteResultDto": {
        "type": "object",
        "properties": { "deleted": { "type": "boolean" } },
        "required": ["deleted"]
      },
...
      "Object": { "type": "object", "properties": {} },
...
      "CollaborationMessageSummaryDto": {
        "type": "object",
        "properties": {
          "total": { "type": "number" },
          "unread": { "type": "number" },
          "read": { "type": "number" },
          "archived": { "type": "number" }
        },
        "required": ["total", "unread", "read", "archived"]
      },
...
      "CollaborationSummaryDto": {
        "type": "object",
        "properties": {
          "messages": {
            "$ref": "#/components/schemas/CollaborationMessageSummaryDto"
          },
          "notices": {
            "$ref": "#/components/schemas/CollaborationNoticeSummaryDto"
          },
          "todos": {
            "$ref": "#/components/schemas/CollaborationTodoSummaryDto"
          },
          "approvals": {
            "$ref": "#/components/schemas/CollaborationApprovalSummaryDto"
          }
        },
        "required": ["messages", "notices", "todos", "approvals"]
      },
      "MessageDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "body": { "type": "string" },
          "sender": { "type": "string" },
          "recipient": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["unread", "read", "archived", "deleted"]
          },
          "businessType": { "type": "string" },
          "businessId": { "type": "string" },
          "readAt": { "type": "string" },
          "archivedAt": { "type": "string" },
          "deletedAt": { "type": "string" },
          "createdAt": { "type": "string" }
        },
        "required": [
          "id",
          "title",
          "body",
          "sender",
          "recipient",
          "status",
          "createdAt"
        ]
      },
      "MessagePageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MessageDto" }
          },
          "page": { "type": "number" },
          "pageSize": { "type": "number" },
          "total": { "type": "number" },
          "totalPages": { "type": "number" }
        },
        "required": ["items", "page", "pageSize", "total", "totalPages"]
      },
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "body": { "type": "string" },
          "sender": { "type": "string" },
          "recipient": { "type": "string" },
          "businessType": { "type": "string" },
          "businessId": { "type": "string" }
        },
        "required": ["title", "body", "sender", "recipient"]
      },
      "NoticeDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "body": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["draft", "published", "archived"]
          },
          "targetAudience": { "type": "array", "items": { "type": "string" } },
          "validFrom": { "type": "string" },
          "validTo": { "type": "string" },
          "publishedAt": { "type": "string" },
          "archivedAt": { "type": "string" },
          "createdBy": { "type": "string" },
          "createdAt": { "type": "string" }
        },
        "required": [
          "id",
          "title",
          "body",
          "status",
          "targetAudience",
          "createdBy",
          "createdAt"
        ]
      },
      "NoticePageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/NoticeDto" }
          },
          "page": { "type": "number" },
          "pageSize": { "type": "number" },
          "total": { "type": "number" },
          "totalPages": { "type": "number" }
        },
        "required": ["items", "page", "pageSize", "total", "totalPages"]
      },
      "CreateNoticeDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "body": { "type": "string" },
          "targetAudience": { "type": "array", "items": { "type": "string" } },
          "createdBy": { "type": "string" },
          "validFrom": { "type": "string" },
          "validTo": { "type": "string" }
        },
        "required": ["title", "body", "targetAudience", "createdBy"]
      },
      "CollaborationTimelineEntryDto": {
        "type": "object",
        "properties": {
          "at": { "type": "string" },
          "actor": { "type": "string" },
          "action": { "type": "string" }
        },
        "required": ["at", "actor", "action"]
      },
      "TodoDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "sourceType": { "type": "string" },
          "businessType": { "type": "string" },
          "businessId": { "type": "string" },
          "assignee": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["pending", "assigned", "completed", "canceled"]
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollaborationTimelineEntryDto"
            }
          },
          "completedAt": { "type": "string" },
          "canceledAt": { "type": "string" },
          "createdAt": { "type": "string" }
        },
        "required": [
          "id",
          "title",
          "sourceType",
          "assignee",
          "status",
          "timeline",
          "createdAt"
        ]
      },
      "TodoPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TodoDto" }
          },
          "page": { "type": "number" },
          "pageSize": { "type": "number" },
          "total": { "type": "number" },
          "totalPages": { "type": "number" }
        },
        "required": ["items", "page", "pageSize", "total", "totalPages"]
      },
      "CreateTodoDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "sourceType": { "type": "string" },
          "businessType": { "type": "string" },
          "businessId": { "type": "string" },
          "assignee": { "type": "string" },
          "actor": { "type": "string" }
        },
        "required": ["title", "sourceType", "assignee", "actor"]
      },
      "AssignTodoDto": {
        "type": "object",
        "properties": {
          "assignee": { "type": "string" },
          "actor": { "type": "string" }
        },
        "required": ["assignee", "actor"]
      },
      "TodoActionDto": {
        "type": "object",
        "properties": { "actor": { "type": "string" } },
        "required": ["actor"]
      },
...
      "ApprovalLitePageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ApprovalLiteDto" }
          },
          "page": { "type": "number" },
          "pageSize": { "type": "number" },
          "total": { "type": "number" },
          "totalPages": { "type": "number" }
        },
        "required": ["items", "page", "pageSize", "total", "totalPages"]
      },
      "CreateApprovalLiteDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "requester": { "type": "string" },
          "approver": { "type": "string" },
          "businessType": { "type": "string" },
          "businessId": { "type": "string" }
        },
        "required": ["title", "requester", "approver"]
      },
      "DecideApprovalLiteDto": {
        "type": "object",
        "properties": {
          "actor": { "type": "string" },
          "comment": { "type": "string" }
        },
        "required": ["actor"]
      },
...
    }
  }
}